+++ nixos/staging.nix
@@ -31,6 +31,7 @@
 
   kpbj.streaming = {
     icecastPort = 8001;
+    icecastHostname = "stream.staging.kpbj.fm";
     webhookPort = 9001;
     apiBase = "http://127.0.0.1:4000/api/playout";
     restartOnDeploy = true;